backend/requirements.txt
# FastAPI Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Dapr SDK
dapr==1.13.0
dapr-ext-grpc==1.13.0

# Anthropic Claude API
anthropic==0.18.1

# Kafka Client
kafka-python==2.0.2

# HTTP Client
httpx==0.26.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.1
python-json-logger==2.0.7

# Prometheus Metrics
prometheus-client==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
